fix(api): make limits consistent with builder pattern and ecosystem
Addresses issues #7 and #8 from api-feedback.md:
- #7: Limits inconsistent with builder pattern
- #8: max_memory is usize instead of u64

Changes:
1. Changed max_memory from usize to u64 for ecosystem consistency
   - zenwebp, zengif, and zencodecs all use u64
   - Prevents silent truncation on 32-bit targets
   - No cast required in adapter code

2. Made max_pixels and max_memory fields pub(crate) instead of pub
   - External users must use builder methods: .max_pixels() / .max_memory()
   - Internal code can still access directly for efficiency
   - Enforces consistent API usage

3. Added getter methods for the now-private fields
   - get_max_pixels() -> u64
   - get_max_memory() -> u64

4. Updated DEFAULT_MAX_MEMORY constant from usize to u64

Before (inconsistent):
```rust
dc.max_pixels = max_px;              // Direct field access
dc.max_memory = max_mem as usize;    // Cast required!
```

After (consistent):
```rust
let dc = DecodeConfig::new()
    .max_pixels(max_px)               // Builder method
    .max_memory(max_mem);             // No cast, u64 -> u64
```

Breaking change: Fields no longer pub (acceptable for 0.x).
Users relying on direct field access must switch to builder methods.

Tests verify:
- Builder methods work correctly
- u64 values don't require casts
- Large limits (5GB) are handled properly
- Fields cannot be accessed directly from external code

+++ zenjpeg/src/foundation/alloc.rs
@@ -39,7 +39,7 @@ pub const MAX_ICC_PROFILE_SIZE: usize = 16 * 1024 * 1024;
 
 /// Default maximum memory for decode operations (512 MB).
 /// This limits total allocations during a single decode operation.
-pub const DEFAULT_MAX_MEMORY: usize = 512 * 1024 * 1024;
+pub const DEFAULT_MAX_MEMORY: u64 = 512 * 1024 * 1024;
 
 /// Tracks cumulative memory allocations during decode operations.
 ///
@@ -66,7 +66,7 @@ impl MemoryTracker {
     /// Creates a new tracker with default limit (512 MB).
     #[must_use]
     pub fn with_default_limit() -> Self {
-        Self::new(DEFAULT_MAX_MEMORY)
+        Self::new(DEFAULT_MAX_MEMORY as usize)
     }
 
     /// Creates an unlimited tracker (for testing or trusted inputs).
